+++ src/js/datepicker.js
@@ -78,6 +78,7 @@
 
             todayButton: false,
             clearButton: false,
+            closeButton: false,
 
             showEvent: 'focus',
             autoClose: false,
@@ -671,6 +672,10 @@
             if (this.opts.todayButton instanceof Date) {
                 this.selectDate(this.opts.todayButton)
             }
+        },
+
+        close: function () {
+            this.hide();
         },
 
         clear: function () {
@@ -1619,6 +1624,7 @@
 		    monthsShort: [ 'Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec' ],
 		    today: 'Today',
 		    clear: 'Clear',
+		    close: 'Close',
 		    dateFormat: 'm/d/Y',
 		    timeFormat: 'h:i a',
 		    firstDay: 0